rename arrowFunction to arrow_function
