+++ .gitignore
@@ -146,3 +146,4 @@ doc-metamodel/*
 # Models
 JSONModels/*
 PUMLModels/*
+.vscode/settings.json
