+++ src/lib/famix/src/model/famix/index.ts
@@ -36,6 +36,6 @@ export { Type } from "./type";
 export { Variable } from "./variable";
 export { ParametricFunction } from "./parametric_function";
 export { ParametricMethod } from "./parametric_method";
-export { ArrowFunction } from "./arrowFunction";
+export { ArrowFunction } from "./arrow_function";
 export { ParametricArrowFunction } from "./parametric_arrow_function";
 export { ParameterConcretisation } from "./parameter_concretisation";
